+++ common/lib/fb.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 struct resolution {
-    uint16_t width;
-    uint16_t height;
+    uint64_t width;
+    uint64_t height;
     uint16_t bpp;
 };
 
@@ -13,9 +13,9 @@ struct fb_info {
     bool default_res;
     uint8_t  memory_model;
     uint64_t framebuffer_addr;
-    uint16_t framebuffer_pitch;
-    uint16_t framebuffer_width;
-    uint16_t framebuffer_height;
+    uint64_t framebuffer_pitch;
+    uint64_t framebuffer_width;
+    uint64_t framebuffer_height;
     uint16_t framebuffer_bpp;
     uint8_t  red_mask_size;
     uint8_t  red_mask_shift;
@@ -26,7 +26,7 @@ struct fb_info {
 };
 
 bool fb_init(struct fb_info *ret,
-             uint16_t target_width, uint16_t target_height, uint16_t target_bpp);
+             uint64_t target_width, uint64_t target_height, uint16_t target_bpp);
 
 void fb_clear(struct fb_info *fb);
